# api-docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "2.5.4",
    "title": "CVE Services API",
...
    "contact": {
      "name": "CVE Services Overview",
      "url": "https://www.cve.org/AllResources/CveServices"
    }
  },
  "servers": [
    {
      "url": "https://cveawg-dev.mitre.org/api"
    }
  ],
  "paths": {
...
    "/registryOrg": {
      "get": {
        "tags": [
          "Registry Organization"
        ],
        "summary": "Retrieves information about all registry organizations (accessible to Secretariat only)",
        "description": "  <h2>Access Control</h2>  <p>Only users with Secretariat role can access this endpoint</p>  <h2>Expected Behavior</h2>  <p><b>Secretariat:</b> Retrieves a list of all registry organizations</p>",
        "operationId": "getAllRegistryOrgs",
        "parameters": [
          {
            "name": "registry",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/pageQuery"
          },
          {
            "$ref": "#/components/parameters/apiEntityHeader"
          },
          {
            "$ref": "#/components/parameters/apiUserHeader"
          },
          {
            "$ref": "#/components/parameters/apiSecretHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of all registry organizations, along with pagination fields if results span multiple pages of data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/registry-org/get-registry-org-response.json"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/bad-request.json"
                }
              }
            }
          },
          "401": {
            "description": "Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Registry Organization"
        ],
        "summary": "Creates a new registry organization (accessible to Secretariat only)",
        "description": "  <h2>Access Control</h2>  <p>Only users with Secretariat role can access this endpoint</p>  <h2>Expected Behavior</h2>  <p><b>Secretariat:</b> Creates a new registry organization</p>",
        "operationId": "createRegistryOrg",
        "parameters": [
          {
            "name": "registry",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/apiEntityHeader"
          },
          {
            "$ref": "#/components/parameters/apiUserHeader"
          },
          {
            "$ref": "#/components/parameters/apiSecretHeader"
          }
        ],
        "responses": {
          "201": {
            "description": "The registry organization was successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/org/create-org-response.json"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/bad-request.json"
                }
              }
            }
          },
          "401": {
            "description": "Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgPayload"
              }
            }
          }
        }
      }
    },
    "/registryOrg/{identifier}": {
      "get": {
        "tags": [
          "Registry Organization"
        ],
        "summary": "Retrieves information about a specific registry organization",
        "description": "  <h2>Access Control</h2>  <p>All authenticated users can access this endpoint</p>  <h2>Expected Behavior</h2>  <p><b>All Users:</b> Retrieves information about the specified registry organization</p>",
        "operationId": "getSingleRegistryOrg",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The identifier of the registry organization"
          },
          {
            "name": "registry",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/apiEntityHeader"
          },
          {
            "$ref": "#/components/parameters/apiUserHeader"
          },
          {
            "$ref": "#/components/parameters/apiSecretHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested registry organization information is returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/org/get-org-response.json"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/bad-request.json"
                }
              }
            }
          },
          "401": {
            "description": "Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Registry Organization"
        ],
        "summary": "Deletes an existing registry organization (accessible to Secretariat only)",
        "description": "  <h2>Access Control</h2>  <p>Only users with Secretariat role can access this endpoint</p>  <h2>Expected Behavior</h2>  <p><b>Secretariat:</b> Deletes an existing registry organization</p>",
        "operationId": "deleteRegistryOrg",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The identifier of the registry organization to delete"
          },
          {
            "name": "registry",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/apiEntityHeader"
          },
          {
            "$ref": "#/components/parameters/apiUserHeader"
          },
          {
            "$ref": "#/components/parameters/apiSecretHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "The registry organization was successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/org/delete-org-response.json"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/bad-request.json"
                }
              }
            }
          },
          "401": {
            "description": "Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          }
        }
      }
    },
    "/registryOrg/{shortname}": {
      "put": {
        "tags": [
          "Registry Organization"
        ],
        "summary": "Updates an existing registry organization (accessible to Secretariat only)",
        "description": "  <h2>Access Control</h2>  <p>Only users with Secretariat role can access this endpoint</p>  <h2>Expected Behavior</h2>  <p><b>Secretariat:</b> Updates an existing registry organization</p>",
        "operationId": "updateRegistryOrg",
        "parameters": [
          {
            "name": "shortname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The Shortname of the registry organization to update"
          },
          {
            "name": "registry",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/apiEntityHeader"
          },
          {
            "$ref": "#/components/parameters/apiUserHeader"
          },
          {
            "$ref": "#/components/parameters/apiSecretHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "The registry organization was successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/org/update-org-response.json"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/bad-request.json"
                }
              }
            }
          },
          "401": {
            "description": "Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/errors/generic.json"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrgPayload"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "apiEntityHeader": {
        "in": "header",
        "name": "CVE-API-ORG",
        "description": "The shortname for the organization associated with the user requesting authentication",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "apiUserHeader": {
        "in": "header",
        "name": "CVE-API-USER",
        "description": "The username for the account making the request",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "apiSecretHeader": {
        "in": "header",
        "name": "CVE-API-KEY",
        "description": "The user's API key",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "pageQuery": {
        "in": "query",
        "name": "page",
        "description": "The current page in the paginator",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1
        }
      },
...
    },
...
  }
}